+++ src/views/dashboards/admin/akun/DaftarAkunMahasiswa/index.ts
@@ -1,0 +1,1 @@
+export { default as DaftarMahasiswa } from "./DaftarAkunMahasiswa";
