+++ src/views/dashboards/admin/akun/DaftarAkunOrganisator/index.ts
@@ -1,0 +1,1 @@
+export { default as DaftarOrganisator } from "./DaftarAkunOrganisator";
