+++ src/views/dashboards/admin/akun/DaftarAkunPanitia/index.ts
@@ -1,0 +1,1 @@
+export { default as DaftarPanitia } from "./DaftarAkunPanitia";
