+++ src/views/dashboards/admin/akun/index.ts
@@ -1,3 +1,6 @@
 export { TambahMahasiswa } from "./TambahAkunMahasiswa";
 export { TambahPanitia } from "./TambahAkunPanitia";
 export { TambahOrganisator } from "./TambahAkunOrganisator";
+export { DaftarMahasiswa } from "./DaftarAkunMahasiswa";
+export { DaftarPanitia } from "./DaftarAkunPanitia";
+export { DaftarOrganisator } from "./DaftarAkunOrganisator";
