--- src/views/dashboards/admin/state/DaftarOrganisator/index.ts
@@ -1,1 +1,0 @@
-export { default as DaftarOrganisator } from './DaftarOrganisator';
