+++ src/views/dashboards/admin/state/index.ts
@@ -1,7 +1,6 @@
-export { TambahState } from './TambahState';
-export { DaftarState } from './DaftarState';
-export { StateDetail } from './StateDetail';
-export { EditState } from './EditState';
-export { DaftarOrganisator } from './DaftarOrganisator';
-export { TambahPIC } from './TambahPIC';
-export { DaftarPIC } from './DaftarPIC';
+export { TambahState } from "./TambahState";
+export { DaftarState } from "./DaftarState";
+export { StateDetail } from "./StateDetail";
+export { EditState } from "./EditState";
+export { TambahPIC } from "./TambahPIC";
+export { DaftarPIC } from "./DaftarPIC";
